refactor: move getSearchTitle to db.ts
